requirements.txt
python-dotenv
openai
PyPDF2
sentence_transformers
langchain-huggingface
langchain-openai
uvicorn
fastapi
langserve[all]
langsmith
pinecone-client
unstructured
unstructured[pdf]
langchain_community
streamlit